fix memory leaks

+++ app/src/main/cpp/WINAPI_KERNEL/SDK/include/Windows/Kernel/WindowsAPIObject.h
@@ -32,7 +32,7 @@ typedef class Object {
         char *name;
         ObjectFlag flags;
         HANDLE handle;
-        int handles;
+        size_t handles;
         ResourceType resource;
 
         void clean();
